backend/requirements.txt
amazon-textract-response-parser==1.0.3
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
bcrypt==4.2.0
bidict==0.23.1
billiard==4.2.0
blinker==1.8.2
boto3==1.35.42
botocore==1.35.42
cachetools==5.3.3
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.3
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.0
fastapi==0.110.0
ffmpeg-python==0.2.0
future==1.0.0
google-api-core==2.17.1
google-auth==2.28.2
google-cloud==0.34.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.15.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.4
jmespath==1.0.1
kombu==5.3.5
lxml==5.1.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mongoengine==0.29.1
numpy==1.26.4
openai==1.13.3
packaging==24.0
passlib==1.7.4
pdf2image==1.17.0
pillow==10.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
PyExecJS==1.5.1
pyinstrument==4.6.2
pymongo==4.10.1
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-jose==3.3.0
python-multipart==0.0.9
python-pptx==0.6.23
python-socketio==5.11.4
requests==2.31.0
rsa==4.9
s3transfer==0.10.3
setuptools==75.1.0
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
tomli==2.0.1
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.28.0
vine==5.1.0
wcwidth==0.2.13
wheel==0.44.0
wsproto==1.2.0
XlsxWriter==3.2.0
